test plant geters

+++ src/action_arroser.cc
@@ -18,7 +18,7 @@ int ActionArroser::check(const GameState& st) const
         return DEJA_ARROSEE;
     if (st.get_map().plant_at(position_plante_)->age < AGE_DE_POUSSE)
         return PAS_ENCORE_ADULTE;
-    if (amelioration_ < 0 || amelioration_ >= 4)
+    if (amelioration_ < 0 || (int)amelioration_ >= 4)
         return CARACTERISTIQUE_INVALIDE;
     return OK;
 }
